Relocated console visibility toggle functions to Terminal component
