+++ src/pages/invocation/invocationPage.tsx
@@ -103,7 +103,6 @@ function InvocationPageContent({
   setLoading: (loading: boolean) => void;
 }>) {
   const { handleUpdateNetwork } = useNetwork(false);
-  const [isTerminalVisible, setIsTerminalVisible] = useState(true);
   const [isModalOpen, setIsModalOpen] = useState(false);
 
   const {
@@ -133,13 +132,6 @@ function InvocationPageContent({
     setIsModalOpen(false);
   }
 
-  function toggleTerminalVisibility(event: KeyboardEvent) {
-    if (event.ctrlKey && event.key === 'j') {
-      event.preventDefault();
-      setIsTerminalVisible((prev) => !prev);
-    }
-  }
-
   function handleLoadContractWithLoading(contractId: string) {
     if (!loading && !isLoadingContract) setLoading(true);
     handleLoadContract(contractId);
@@ -148,14 +140,6 @@ function InvocationPageContent({
   useEffect(() => {
     if (data.contractId && loading) setLoading(false);
   }, [data, loading, setLoading]);
-
-  useEffect(() => {
-    window.addEventListener('keydown', toggleTerminalVisibility);
-
-    return () => {
-      window.removeEventListener('keydown', toggleTerminalVisibility);
-    };
-  }, []);
 
   return (
     <Fragment>
@@ -190,9 +174,8 @@ function InvocationPageContent({
               data={data}
               preInvocationValue={preInvocationValue}
               postInvocationValue={postInvocationValue}
-              setIsTerminalVisible={setIsTerminalVisible}
             />
-            {isTerminalVisible && <Terminal entries={contractResponses} />}
+            <Terminal entries={contractResponses} />
           </div>
         ) : (
           <InvocationCTAPage />
